Create admin page - delete graduated student
Create page that displays a list of all the students´ usernames and ids so that the admin can delete the graduated students from the database by providing the username and id. Also deleted extra files not needed for our project. For this two routes were created - a get route in students routes to get the list of students and a delete route to delete the selected one from the database